pyproject.toml
[project]
name = "mtg-deck-builder"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.3.34",
    "langchain-openai>=0.3.12",
    "langchain-ollama",
    "langchain-postgres",
    "pgvector",
    "psycopg[binary]",
    "sqlalchemy",
    "pydantic>=2.6",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "python-dotenv",
    "tenacity",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mtg"]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
